agent-producer/requirements.txt
confluent-kafka
python-dotenv
winsdk
